+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `modifier` prop to `HorizontalStroke`.
+- Feature: Add `SingleEntryTable` component.
+Feature:
+- Add `ExportIconAlternate` icon.
+- Feature: Add `margin` and `align` props to `Pagination`.
 
 ## [2.68.0] - 2020-05-06